LayoutTests/fast/forms/ValidityState-customError-expected.txt

This test checks validity.customError.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Set some value for customError.
PASS customErrorFor("fieldset-set") is false
PASS customErrorFor("button-set") is true
PASS customErrorFor("button-button-set") is false
PASS customErrorFor("button-reset-set") is false
PASS customErrorFor("select-set") is true
PASS customErrorFor("textarea-set") is true
PASS customErrorFor("input-set") is true
PASS customErrorFor("input-submit-set") is true

Set some value for customError, and clear it.
PASS customErrorFor("fieldset-set-and-empty") is false
PASS customErrorFor("button-set-and-empty") is false
PASS customErrorFor("button-button-set-and-empty") is false
PASS customErrorFor("button-reset-set-and-empty") is false
PASS customErrorFor("select-set-and-empty") is false
PASS customErrorFor("textarea-set-and-empty") is false
PASS customErrorFor("input-set-and-empty") is false
PASS customErrorFor("input-submit-set-and-empty") is false

Not any changes for customError.
PASS customErrorFor("fieldset-no-changes") is false
PASS customErrorFor("button-no-changes") is false
PASS customErrorFor("button-button-no-changes") is false
PASS customErrorFor("button-reset-no-changes") is false
PASS customErrorFor("select-no-changes") is false
PASS customErrorFor("textarea-no-changes") is false
PASS customErrorFor("input-no-changes") is false
PASS customErrorFor("input-submit-no-changes") is false

Many changes for customError:
Set some value, and set with no arguments.
PASS setCustomValidity() threw Not enough arguments
PASS setCustomValidity() threw Not enough arguments
PASS setCustomValidity() threw Not enough arguments
PASS setCustomValidity() threw Not enough arguments
PASS setCustomValidity() threw Not enough arguments
PASS setCustomValidity() threw Not enough arguments
PASS setCustomValidity() threw Not enough arguments
PASS setCustomValidity() threw Not enough arguments
PASS customErrorFor("fieldset-many-changes") is false
PASS customErrorFor("button-many-changes") is true
PASS customErrorFor("button-button-many-changes") is false
PASS customErrorFor("button-reset-many-changes") is false
PASS customErrorFor("select-many-changes") is true
PASS customErrorFor("textarea-many-changes") is true
PASS customErrorFor("input-many-changes") is true
PASS customErrorFor("input-submit-many-changes") is true
Set with three arguments.
PASS customErrorFor("fieldset-many-changes") is false
PASS customErrorFor("button-many-changes") is true
PASS customErrorFor("button-button-many-changes") is false
PASS customErrorFor("button-reset-many-changes") is false
PASS customErrorFor("select-many-changes") is true
PASS customErrorFor("textarea-many-changes") is true
PASS customErrorFor("input-many-changes") is true
PASS customErrorFor("input-submit-many-changes") is true
Set null.
PASS customErrorFor("fieldset-many-changes") is false
PASS customErrorFor("button-many-changes") is true
PASS customErrorFor("button-button-many-changes") is false
PASS customErrorFor("button-reset-many-changes") is false
PASS customErrorFor("select-many-changes") is true
PASS customErrorFor("textarea-many-changes") is true
PASS customErrorFor("input-many-changes") is true
PASS customErrorFor("input-submit-many-changes") is true
Set some value, and set undefined.
PASS customErrorFor("fieldset-many-changes") is false
PASS customErrorFor("button-many-changes") is true
PASS customErrorFor("button-button-many-changes") is false
PASS customErrorFor("button-reset-many-changes") is false
PASS customErrorFor("select-many-changes") is true
PASS customErrorFor("textarea-many-changes") is true
PASS customErrorFor("input-many-changes") is true
PASS customErrorFor("input-submit-many-changes") is true

PASS successfullyParsed is true

TEST COMPLETE
